+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - test eccezione generica
 
+### Changed
+
+- suddivisione test normali (DocResourceTest) / test sicurezza (DocResourceSicurezzaTest)
+
 ## [1.0.0] - 2025-11-12
 
 ### Added